feat(generator): adds highlighting
In order to do so, it also provides add two modes in the configuration
editor: select and highlight. The select mode is activated when one
of the css selectors button is activated, in this mode the selector UI
is active. The highlighted is activated otherwise, and displays the
different matching zone with colors.