Add minimal support for Javascript-rendered documentations
This commit adds a minimal support for Javascript-generated documentations. It
makes it possible by using ScrapyHub's Splash as a proxy to render those
documentations before crawling them.

On the user-side, the `js_render` boolean parameter is used to enable the
proxy connection in the configuration file. A Splash instance is also needed to
serve as a proxy (see the README.md to know how to start it using Docker).

This commit also adds the configuration files for the following documentations:

 - Prezly.com
 - Replicated.com