Integrate a Splash instance in the Docker container
This commit adds a Splash instance in our Docker containers. The Splash
instance is run only if needed i.e. if the JSON configuration file has the
`js_render` parameter set to true.

On the long term, we should considerate to separate those two instances and to
run them properly on the Mesos cluster.