Split the Docker image to reduce build time
Since we've added the Splash instance to our Docker image, the build time has
increased a lot because of all the new dependencies it relies on.

This commit splits the Docker image by creating a `base` image containing the
Splash instance and all its dependencies. The Docker image that we'll need to
build everyday is just adding our `documentation-scrapper` code.