config(reindex): Adding reindex.io config
The sidebar menu can have up to two levels. What the `lvl0` selector
does is first trying to get the currently selected page, and if it is
a subpage then grab its parent category. If that fails, it will get
the currently selected link. I'm not searching into that field.

Other selectors are classical h1/h2. For pages that are at the top of
the hierarchy, the `lvl0` and `lvl1` will be identical.