Replace Splash with PhantomJS as a proxy for JS documentations
We used to rely on Splash to handle JS documentations by using it as proxy for
every request in order to render the HTML code before to crawl it. However, it
doesn't let us handle corner cases properly.

This commit replaces the Splash instance with PhantomJS. It lets us to have a
better fine-grained control on how we handle requests for JS documentations.