test(min_indexed_level): Add test for `min_indexed_level`.
I've added a simple test that checks that for a defined value of
`min_indexed_level`, the `n` first levels are correctly not indexed.

I've also added a few Docstring to some tests where it was missing.

The `get_records_from_dom` is a huge monster. We'll have to split it
one day, it is getting more and more complex to understand (I'm sure
we could simplify it by learning a bit more of python, I can't believe
the language forces us all those loops).

Fixes #29