deployer: ConfigCreator: extract_root_from_input: Remove useless backslashes
In Python, escape sequences (backslash followed by a character) that
are unknown are tolerated, and are treated as if the backslash was a
plain character (instead of a control one) - as if it was doubled.

So "\/" and "\\/" contains the same characters (backslash followed by
slash).

On the other hand, in Python, in a regular expression pattern, escape
seauences (backslash followed by a characted) that are unknwon are
tolerated, and treated as if the backslash was not there.

So "\/" and "/" as string regex patterns match the same thing, a single
slash.

as "/" is no control character. It works at the moment because Python
tolerates invalid/unknown escape sequences by ignoring the backslash.