Update algoliasearch to 2.6.3
It is currently not possible to run Algolia search tests in the same script that generates a test Algolia index.
The reason is we pull in v1.13 here and that conflicts with the v2.6.3 we need to pull in applied2 when running
our tests using the Algolia Python API.

This PR updates algoliasearch to match the version we are using in applied2.